+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for `up` and `down` direction for piston methods
+
 ## [0.3.0] - 05.05.22
 
 ### Added